added verification for member area

+++ protected/modules/member/controllers/AlbumController.php
@@ -6,12 +6,11 @@ class AlbumController extends YsaMemberController
 		
 		$this->crumb('Events', array('event/'));
 	}
-
-
+	
 	public function actionCreate($event)
     {
-        $event = Event::model()->findByPk($event);
-        
+        $event = Event::model()->findByIdLogged($event);
+		
         // no event found or it's a proofing event
         if (!$event || Event::TYPE_PROOF == $event->type) {
             $this->redirect(array('event/'));
@@ -69,7 +68,7 @@ class AlbumController extends YsaMemberController
     {
         $entry = EventAlbum::model()->findByPk($albumId);
         
-        if (!$entry || !$entry->event()) {
+        if (!$entry || !$entry->event() || !$entry->event()->isOwner()) {
             $this->redirect(array('event/'));
         }
 		$upload = new PhotoUploadForm();
@@ -92,6 +91,8 @@ class AlbumController extends YsaMemberController
 		$this->crumb($entry->event()->name, array('event/view/' . $entry->event()->id))
 			 ->crumb($entry->name);
 		
+		$this->setMemberPageTitle($entry->name);
+		
         $this->render('view', array(
             'entry'   => $entry,
 			'upload'  => $upload, 
@@ -111,7 +112,9 @@ class AlbumController extends YsaMemberController
 			$album = EventAlbum::model()->findByPk($id);
 			if ($album) {
 				$event = $album->event();
-				$album->delete();	
+				if ($event->isOwner()) {
+					$album->delete();
+				}
 			}
         }
         
@@ -132,7 +135,7 @@ class AlbumController extends YsaMemberController
 			if (isset($_POST['event-album'])) {
 				foreach ($_POST['event-album'] as $k => $id) {
 					$entry = EventAlbum::model()->findByPk($id);
-					if ($entry) {
+					if ($entry && $entry->event()->isOwner()) {
 						$entry->rank = $k + 1;
 						$entry->save();
 					}
